% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_integration.R
\name{ggplot_add.vbracket_legend}
\alias{ggplot_add.vbracket_legend}
\title{Add vbracket legend to ggplot}
\usage{
\method{ggplot_add}{vbracket_legend}(object, plot, ...)
}
\arguments{
\item{object}{A vbracket_legend object}

\item{plot}{A ggplot object}

\item{...}{Additional arguments (not used)}
}
\description{
Add vbracket legend to ggplot
}
