% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_integration.R
\name{ggplot_build.gg}
\alias{ggplot_build.gg}
\title{Plot ggplot with vbracket legend}
\usage{
\method{ggplot_build}{gg}(x)
}
\arguments{
\item{x}{A ggplot object}
}
\description{
Plot ggplot with vbracket legend
}
\keyword{internal}
