% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_integration.R
\name{ggplot_gtable.gg}
\alias{ggplot_gtable.gg}
\title{Print method for ggplot with vbracket legend}
\usage{
\method{ggplot_gtable}{gg}(x)
}
\arguments{
\item{x}{A ggplot object with vbracket legend}

\item{newpage}{Draw on new page}

\item{vp}{Viewport}

\item{...}{Additional arguments}
}
\description{
Print method for ggplot with vbracket legend
}
\keyword{internal}
