% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_integration.R
\name{print_vbracket_plot}
\alias{print_vbracket_plot}
\title{Print a ggplot with vbracket overlay}
\usage{
print_vbracket_plot(x, ...)
}
\arguments{
\item{x}{ggplot object}

\item{...}{Additional arguments}
}
\description{
Print a ggplot with vbracket overlay
}
\keyword{internal}
