% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_integration.R
\name{ggplot_add.vbracket_legend}
\alias{ggplot_add.vbracket_legend}
\title{Add vbracket legend to ggplot}
\usage{
\method{ggplot_add}{vbracket_legend}(object, plot, object_name)
}
\arguments{
\item{object}{A ggplot object}

\item{plot}{A vbracket_legend object}

\item{object_name}{Not used}
}
\description{
Add vbracket legend to ggplot
}
