% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_method.R
\name{print.gg}
\alias{print.gg}
\title{Print method for ggplot objects with vbracket legend}
\usage{
\method{print}{gg}(x, newpage = is.null(vp), vp = NULL, ...)
}
\arguments{
\item{x}{A ggplot object}

\item{newpage}{Draw on new page (default TRUE)}

\item{vp}{Viewport to draw in}

\item{...}{Additional arguments}
}
\description{
This overrides the default ggplot print method to add brackets automatically
}
